refactor: Update axios-instance.ts to include fetcher functions
This commit updates the axios-instance.ts file to include two new fetcher functions: fetcherWithoutData and fetcher. The fetcherWithoutData function is used to make a GET request and return the response data, while the fetcher function is used to make a GET request with additional configurations and return either the response data or the entire response object. These new functions provide more flexibility and customization when making API requests using axios.